Yoni - changed some structure
